Easily switch floaterm and files + resize floater
Switch to the left window with C-h even on floaterm (TERIMNAL mode), C-i
to switch to the above window (C-k is already used by fzf), C-space to
use a terminal in normal mode (primarily to yank text)

Resize floaterm on the fly with Ctrl+arrow keys. The horizontal resize
is slower.