nvim restore lastline simpler, vv alias no set var
The . mark contains the last line. Execute that command silently to
hide error messages if that line was removed with another editor instead
of checking with ifs

There's no need to set the plug variable using the vv alias. With native
plugins the colorscheme works even without loading vim-colorschemes, and
--noplugin also prevents pak/.. plugins from being loaded